move funcs from test/common.py to (new) sim_if/common.py
