+++ vunit/builtins.py
@@ -8,11 +8,10 @@
 Functions to add builtin VHDL code to a project for compilation
 """
 
-
 from os.path import join, abspath, dirname, basename
 from glob import glob
 from vunit.vhdl_standard import VHDL
-from vunit.test.common import simulator_check
+from vunit.sim_if.common import simulator_check
 
 VHDL_PATH = abspath(join(dirname(__file__), "vhdl"))
 VERILOG_PATH = abspath(join(dirname(__file__), "verilog"))
